Added new _recordreader file and updated paths
